requirements/common.txt
Flask==0.10.1
Flask-Bootstrap==3.0.3.1
Flask-Mail==0.9.0
Flask-Migrate==1.1.0
Flask-Moment==0.2.1
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Flask-WTF==0.9.4
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
SQLAlchemy==0.9.9
WTForms==1.0.5
Werkzeug==0.10.4
alembic==0.6.2
blinker==1.3
itsdangerous==0.23

flask-login==0.3.2
flask-pagedown==0.2.2
markdown==2.6.7
bleach==1.4.3

# flask rest 服务
flask-httpauth==3.2.1

# mongodb
flask-pymongo==0.4.1